pyproject.toml
[tool.black]
line-length = 79
target-version = ['py311']  # Set the Python versions you want to target
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "sudoku-solver"
version = "0.1.0"
description = "Sudoku solver to learn Python API and DB manipulation"
authors = ["Agata <tekienagata@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
flask = "^2.3.3"
flask-restx = "^1.2.0"
flask_sqlalchemy = "^3.1.1"
pydantic = "^2.6.1"
psycopg2-binary = "^2.9.9"
testcontainers = "^4.3.1"
docker = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0 "
flake8 = "^6.1.0"
isort = "^5.13.2"
pylint = "^3.0.3"

[tool.pylint."messages_control"]
disable = [
    "W0621"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"